requirements.txt
textblob==0.17.1
tweepy==4.14.0
python-dotenv==1.0.0